updated the endpoints for auth tokens

+++ application/endpoints/users.py
@@ -8,6 +8,7 @@ from pydantic.networks import EmailStr
 from sqlalchemy.orm import Session
 from tempfile import NamedTemporaryFile
 import os
+from typing import Union
 
 import boto3, botocore
 import crud, model as models, schemas
@@ -94,28 +95,28 @@ def create_user(
     return user #json.dumps({"user":{"email":user.email, "full_name":user.full_name}, "token":access_token})
 
 
-@router.put("/me", response_model=schemas.User)
-def update_user_me(
-    *,
-    db: Session = Depends(deps.get_db),
-    password: str = Body(None),
-    full_name: str = Body(None),
-    email: EmailStr = Body(None),
-    current_user: models.User = Depends(deps.get_current_active_user),
-) -> Any:
-    """
-    Update own user.
-    """
-    current_user_data = jsonable_encoder(current_user)
-    user_in = schemas.UserUpdate(**current_user_data)
-    if password is not None:
-        user_in.password = password
-    if full_name is not None:
-        user_in.full_name = full_name
-    if email is not None:
-        user_in.email = email
-    user = crud.user.update(db, db_obj=current_user, obj_in=user_in)
-    return user
+# @router.put("/me", response_model=schemas.User)
+# def update_user_me(
+#     *,
+#     db: Session = Depends(deps.get_db),
+#     password: str = Body(None),
+#     full_name: str = Body(None),
+#     email: EmailStr = Body(None),
+#     current_user: models.User = Depends(deps.get_current_active_user),
+# ) -> Any:
+#     """
+#     Update own user.
+#     """
+#     current_user_data = jsonable_encoder(current_user)
+#     user_in = schemas.UserUpdate(**current_user_data)
+#     if password is not None:
+#         user_in.password = password
+#     if full_name is not None:
+#         user_in.full_name = full_name
+#     if email is not None:
+#         user_in.email = email
+#     user = crud.user.update(db, db_obj=current_user, obj_in=user_in)
+#     return user
 
 
 @router.get("/me", response_model=schemas.User)
@@ -155,11 +156,12 @@ def read_user_by_id(
 def update_userProfile(
     *,
     db: Session = Depends(deps.get_db),
-    phone: str = Form(),
-    address: str = Form(),
-    location: str = Form(),
-    hospital: str = Form(),
-    userProfile: UploadFile = File(...),
+    phone: str = Form(default=None),
+    address: str = Form(default=None),
+    location: str = Form(default=None),
+    # hospital: str = Form(default=None),
+    userProfile: Union[UploadFile, None] = File(default=None),
+    # userProfile: UploadFile = File(default=None),
     current_user: models.User = Depends(deps.get_current_active_user),
 ) -> Any:
     """
@@ -172,42 +174,35 @@ def update_userProfile(
             status_code=404,
             detail="The user with this username does not exist in the system",
         )
-
-    temp = NamedTemporaryFile(delete=False)
-    try:
+    if userProfile:
+        temp = NamedTemporaryFile(delete=False)
         try:
-            contents = userProfile.file#.read()
-            # with temp as f:
-            #     f.write(contents);
-        except Exception:
-            return {"message": "There was an error uploading the file"}
+            try:
+                contents = userProfile.file#.read()
+                # with temp as f:
+                #     f.write(contents);
+            except Exception:
+                return {"message": "There was an error uploading the file"}
+            finally:
+                # Here, upload the file to your S3 service using `temp.name`
+                s3.upload_fileobj(
+                userProfile.file,
+                os.getenv("AWS_BUCKET_NAME"),
+                '%s/%s/%s' % (f"{os.getenv('AWS_BUCKET_FOLDER')}", os.getenv('AWS_PROFILE_FOLDER'),f"{user.full_name}_{userProfile.filename}"))
+                userProfile.file.close()
+        except Exception as e:
+            return {"message": f"There was an error processing the file  as {e}"}
         finally:
-
-            # Here, upload the file to your S3 service using `temp.name`
-
-            #get the file extension
-
-
-            s3.upload_fileobj(
-            userProfile.file,
-            os.getenv("AWS_BUCKET_NAME"),
-            '%s/%s/%s' % (f"{os.getenv('AWS_BUCKET_FOLDER')}", os.getenv('AWS_PROFILE_FOLDER'),f"{user.full_name}_{userProfile.filename}"))
-            userProfile.file.close()
-    except Exception as e:
-        return {"message": f"There was an error processing the file  as {e}"}
-    finally:
-        #temp.close()  # the `with` statement above takes care of closing the file
-        os.remove(temp.name)  # Delete temp file
-
-    # print(contents)  # Handle file contents as desired
-
-    file_uploaded_path = '%s/%s/%s' % (f"{os.getenv('AWS_BUCKET_FOLDER')}", os.getenv('AWS_PROFILE_FOLDER'),f"{user.full_name}_{userProfile.filename}")
-    print("File Name is ", file_uploaded_path)
+            #temp.close()  # the `with` statement above takes care of closing the file
+            os.remove(temp.name)  # Delete temp file
+        # print(contents)  # Handle file contents as desired
+        file_uploaded_path = '%s/%s/%s' % (f"{os.getenv('AWS_BUCKET_FOLDER')}", os.getenv('AWS_PROFILE_FOLDER'),f"{user.full_name}_{userProfile.filename}")
+    else:
+        file_uploaded_path = user.userProfile
     update_data = {
     "phone": phone if phone else user.phone,
     "address": address if address else user.address,
     "location": location if location else user.location,
-    "hospital": hospital if hospital else user.hospital,
     "userProfile": file_uploaded_path if userProfile else user.userProfile,
     }
     user = crud.user.update(db, db_obj=user, obj_in=update_data)
